+++ test/dummy/prepare_cli_worker.rb
@@ -6,10 +6,12 @@ require_relative "config/environment"
 require_relative "../../db/migrate/20260805000000_create_solid_objects_tables"
 require_relative "../../db/migrate/20260806000000_add_state_revision_to_solid_objects_instances"
 require_relative "../../db/migrate/20260813000000_rename_message_dispatch_columns"
+require_relative "../../db/migrate/20260915000000_add_solid_objects_effect_recoveries"
 
 CreateSolidObjectsTables.new.migrate(:up)
 AddStateRevisionToSolidObjectsInstances.new.migrate(:up)
 RenameMessageDispatchColumns.new.migrate(:up)
+AddSolidObjectsEffectRecoveries.new.migrate(:up)
 
 now = Time.current
 instance = SolidObjects::Instance.create!(